+++ ICD.Connect.Routing.Mock/Midpoint/MockRouteMidpointControl.cs
@@ -1,16 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 using ICD.Common.Properties;
 using ICD.Common.Utils.Extensions;
-using ICD.Common.Utils.Services;
 using ICD.Connect.API.Commands;
 using ICD.Connect.Devices;
 using ICD.Connect.Routing.Connections;
 using ICD.Connect.Routing.Controls;
-using ICD.Connect.Routing.Endpoints;
 using ICD.Connect.Routing.EventArguments;
-using ICD.Connect.Routing.RoutingGraphs;
 using ICD.Connect.Routing.Utils;
 
 namespace ICD.Connect.Routing.Mock.Midpoint
@@ -39,15 +35,7 @@ namespace ICD.Connect.Routing.Mock.Midpoint
 
 		private readonly SwitcherCache m_Cache;
 
-		private IRoutingGraph m_CachedRoutingGraph;
-
-		/// <summary>
-		/// Gets the routing graph.
-		/// </summary>
-		public IRoutingGraph RoutingGraph
-		{
-			get { return m_CachedRoutingGraph = m_CachedRoutingGraph ?? ServiceProvider.GetService<IRoutingGraph>(); }
-		}
+		private readonly RoutingGraphMidpointConnectionComponent m_MidpointComponent;
 
 		/// <summary>
 		/// Constructor.
@@ -58,6 +46,7 @@ namespace ICD.Connect.Routing.Mock.Midpoint
 			: base(parent, id)
 		{
 			m_Cache = new SwitcherCache();
+			m_MidpointComponent = new RoutingGraphMidpointConnectionComponent(this);
 			Subscribe(m_Cache);
 		}
 
@@ -85,11 +74,7 @@ namespace ICD.Connect.Routing.Mock.Midpoint
 		/// <returns></returns>
 		public override ConnectorInfo GetOutput(int address)
 		{
-			Connection connection = RoutingGraph.Connections.GetOutputConnection(new EndpointInfo(Parent.Id, Id, address));
-			if (connection == null)
-				throw new ArgumentOutOfRangeException("address");
-
-			return new ConnectorInfo(connection.Source.Address, connection.ConnectionType);
+			return m_MidpointComponent.GetOutput(address);
 		}
 
 		/// <summary>
@@ -99,7 +84,7 @@ namespace ICD.Connect.Routing.Mock.Midpoint
 		/// <returns></returns>
 		public override bool ContainsOutput(int output)
 		{
-			return RoutingGraph.Connections.GetOutputConnection(new EndpointInfo(Parent.Id, Id, output)) != null;
+			return m_MidpointComponent.ContainsOutput(output);
 		}
 
 		/// <summary>
@@ -108,9 +93,7 @@ namespace ICD.Connect.Routing.Mock.Midpoint
 		/// <returns></returns>
 		public override IEnumerable<ConnectorInfo> GetOutputs()
 		{
-			return RoutingGraph.Connections
-							   .GetOutputConnections(Parent.Id, Id)
-							   .Select(c => new ConnectorInfo(c.Source.Address, c.ConnectionType));
+			return m_MidpointComponent.GetOutputs();
 		}
 
 		/// <summary>
@@ -143,11 +126,7 @@ namespace ICD.Connect.Routing.Mock.Midpoint
 		/// <returns></returns>
 		public override ConnectorInfo GetInput(int input)
 		{
-			Connection connection = RoutingGraph.Connections.GetInputConnection(new EndpointInfo(Parent.Id, Id, input));
-			if (connection == null)
-				throw new ArgumentOutOfRangeException("input");
-
-			return new ConnectorInfo(connection.Destination.Address, connection.ConnectionType);
+			return m_MidpointComponent.GetInput(input);
 		}
 
 		/// <summary>
@@ -157,7 +136,7 @@ namespace ICD.Connect.Routing.Mock.Midpoint
 		/// <returns></returns>
 		public override bool ContainsInput(int input)
 		{
-			return RoutingGraph.Connections.GetInputConnection(new EndpointInfo(Parent.Id, Id, input)) != null;
+			return m_MidpointComponent.ContainsInput(input);
 		}
 
 		/// <summary>
@@ -166,9 +145,7 @@ namespace ICD.Connect.Routing.Mock.Midpoint
 		/// <returns></returns>
 		public override IEnumerable<ConnectorInfo> GetInputs()
 		{
-			return RoutingGraph.Connections
-							   .GetInputConnections(Parent.Id, Id)
-							   .Select(c => new ConnectorInfo(c.Destination.Address, c.ConnectionType));
+			return m_MidpointComponent.GetInputs();
 		}
 
 		/// <summary>